Remove 'display map option under adaptation tab (#10)
* fix: Remove adaptation tab display map option

Remove the Display Map option on the Economic & Non-Economic tab and Adaptation subtab section.
Functionality remains in case the map generation is added later on.

- Remove adaptation display map button.
- Display chart view control is automatically selected after selecting the Adaptation subtab.

* fix: Remove adaptation tab display map option

Remove the Display Map option on the Economic & Non-Economic tab and Adaptation subtab section.
Functionality remains in case the map generation is added later on.

- Remove adaptation display map button.
- Display chart view control is automatically selected after selecting the Adaptation subtab.